Group IReplacer for easier customisation
Closes #107
